use cgiurl_graph config in dynazoom.html
dynazoom.html uses a hardcoded path even though the master
configuration allows specifying this path as cgiurl_graph

This means that dynazoom.html should be templatized as well, but it's not an as
small change as the one Hiroaki Nakamura kindely provides.

So, commiting his patch as a temporary workaround.

T: Closes #1204