node: only use basename for $0
kjetilho noticed on IRC that :
---
In Ubuntu Precise, the behaviour of setting $0 has changed. If you look in
/proc/PID/stat, you will see the given value of $0 exactly.

In older distros (kernels?) it would remove the directory part implicitly.
I'm not sure if this happens in Perl or in the kernel.

But as that field is limited to 15 characters, when $0 = "/usr/sbin/munin-node",
what will appear is just "/usr/sbin/munin" which means check_procs -C munin-node
in Nagios will fail, pgrep munin-node will fail and so on.
---

This commit reverts to the old behavior : "remove the directory part implicitly".