plugins/Pgsql.pm: allow multiple graph types
This feature extends module flexibility by allowing user to specify certain
graph type in 'graphdraw' parameter for each counter. If you specify an array of
graph types in the form below (assuming you have three counters):

    graphdraw => [ AREA, LINE1, LINE1 ]

Then graph type for each of the counters will be set according to the elements
of the array.  Graph types elements need to be specified in the order or being returned
by 'configquery'.