dists: remove all dists specific stuff
Since the move to git, it's much more practical to have an official repository
as "clean" as possible.

Therefore, I'm putting all packaging specifics outside the repo, explaining
where to find it. This only makes official what happened since a while, as the
dists/ dir is no accurate anymore.

Note that we work even more closely with packagers than before. This enables a
much faster integration and factoring of patches, as it factors out much of the
noise.