plugins/bind9_rndc: fix if named.stats is on tmpfs
In a situation where bind9 named.stats is on tmpfs, the file wont exist after a
reboot until 'rndc stats' is run. Since the plugin opens the file prior to
running the command, the plugin will exit with a file not found error.

Patch simply tests if the file can be read, if not, issues an extra 'rndc
stats' before opening.