plugins/cpuspeed: fix if cpus are at maximum speed
According to janning <vygen@kicktipp.de>:

-------
If you are running your cpus at maximum speed (like with governor performance
in /etc/default/cpufrequtils) the calculation fails. The plugin counts cpu
cycles. It then calculates the difference to the last mesured cycles dividing
by the 5 minutes (as munin is called every 5 minutes).  thats fine BUT, the
plugin is configured to have a maximum value which is the same as the maximum
top frequency of your cpus.  If you are running your cpu at top speed the
calculated value is hitting the max value, maybe just because the time
calculation is not accurate. if the measurement contains 5 minutes and 2
seconds but the calculation is based on 5 minutes, you have to many cpu cycles.
munin does not show values beyond max_value. So you don't see anything in your
graph.

Just adding extra 10% to max_value as a safety area should solve the issue.
-------

The fix implemented has the same logic, but is pure /bin/sh

Closes: D#615957