Add test for POD Spelling errors
This test is disabled by default.

To enable this test, you need to install the perl module Test::Spelling, and
set the environment variable TEST_POD_SPELLING.  You also need to have a
working spell checker Test::Spelling can use, preferably one which matches the
language of the text in the POD sections.