procsys: implement a new plugin to replace entropy, open_files and open_inodes.
This new plugin uses the new Framework to create graphs for all the
sysctl-related data as a signle round-trip.

There is only one change compared to the previous plugins: open_inodes
produces a more meaningful graph of allocated and free inodes, rather
than the previous one that used a fake max value (as it was actually
the kernel's allocation).

Both open_files and open_inodes graphs now use areastacks instead of lines.