Treat empty query param as absent.
With this fix, both
	dynazoom.html?cgiurl_graph=&plugin_name=...
and
	dynazoom.html?plugin_name=
now fall back to the default value for cgiurl_graph.

This is needed to make fallback to the default path work if the user
hasn't specified it explicitly in /etc/munin.conf. Previously, the
empty string would have been used as path for cgiurl_graph.