plugins: start implementing a new framework for plugins writing.
Munin::Plugin::Framework allows a much easier way to write new plugins
by abstracting the underlying protocol between node and master.

The idea is to have the plugins only declare what they want to have
displayed, eventually merging the code for configuration and
value-setting, and leaving it to the framework to generate the
configuration and fetch output.