asyncd: use the same rules as munin-update
munin-update does some squashing on invalid chars when using non-multigraph
plugins. This is done transparently when using the plain old TCP transport.

As munin-async(d) uses the spoolfetch transport that is multigraph-based, it
has the side effect not squashing the root plugin name. This patch squashes the
root plugin name in the same manner as done by munin-update.

It therefore enables to insert the munin-async(d) couple in-place to benefit
transparently.

Closes D:710529