plugins/df_abs: fix the naming for regular devices
As stated by Christoph Biedl:

Commit a4eb420 was created in the intention to append the mountpoint to
non-device mounts like tmpfs in order to make them unique.

However, the quoting in

+               "/*" )

disabled globbing so this clause will never be hit¹. As a result, the
mountpoint is appended to *all* devices like in

| _dev_vda1__boot.value 19868
| _dev_vda5__.value 165710

This field name change will also break the historic data as the rrd file name
on the master changes accordingly.

It might be an idea to stay that way (I consider that a bad one) but at the
moment the plugin's behaviour matches neither the documention nor the one of
the other df* plugins.

¹ Unless somebody naughty creates a device with literally that name.

Closes: D#770745