This corrects the truncating of negative values in the diskstats plugin and calculates its value based on the evidence that this is a overflow value.
This patch was created by ksmadsen in
http://munin-monitoring.org/ticket/1275

The description provided was:

I've traced this problem to the fact that on the RH kernel we are using,
the rd_ticks/wr_ticks in /sys/block/*/stat are 32-bit numbers, so when
the ticks wraps, a negative value is reported.

Looking at the kernel source code, the tick counters are internally all
64-bit, but when they are read from the stat file, the output is limited
to 32-bits.
(cherry picked from commit 91a9e62b14957cabcd8fcfceeb6f656b95eb2261)