master: don't abort node collection on bad plugin
As noted by Christoph Biedl in the ticket :

Using "multigraph" as a multigraph service name is not accepted by
Munin::Master::Node. The check however dies (in config) or croaks (in fetch),
causing in the connection being dropped and no data gathered for that node.

That's a bit harsh, this commit lowers that situation to two error messages.

Note that impact is rather small as this affects the data for that node itself.
A malicious plugin still does harm, but if an evil guy is able to control which
plugins are run on a node, no longer collecting munin data for that one is not
one of the bigger concerns.

[ Thx Christoph for also providing the patch ]

Closes: #1397, C:CVE-2013-6359