p/ipmi_sensor_: fix the environ() usage
As Gabriel Filion reported on d:728840 :

----------------------------------------------------------------------
with version 2.0.6-4 of munin-plugins-extra, I get the following stack
trace when running the ipmi_sensor_ script:

    munin-run ipmi_sensor_u_degrees_c
    Traceback (most recent call last):
      File "/etc/munin/plugins/ipmi_sensor_u_degrees_c", line 73, in <module>
        CACHEDIR = os.environ['MUNIN_PLUGSTATE']
    NameError: name 'os' is not defined

it would seem the script imports different things for the os library
with "from os import xyz" but does not import the library directly. so
near the beginning when it tries to get a variable from then environment,
if crashes.
----------------------------------------------------------------------

Closes: D:728840