Fix thresholds for users containing a dash in their name
As the dash is illegal in the cdef used by rrdtool, the plugin uses
clean_fieldname() on the user name.

Though, the hash containing the mapping between user name and quota
values is keyed by the original user name, yielding an empty value.

This change introduces a new variable that contains the cleaned user
name, and uses the original user name it to look up values in the
quota hash.