Graph system peer's stratum
Although not ideal to dual purpose the state axis as stratum for this field, I don't think it warrants a separate plugin. Also, it is only enabled if env.show_syspeer_stratum is set (similar to env.lowercase). Especially with the use of NTP pools, this also allows for warn/crit levels if the esystem peer's stratum becomes above an expected stratum.