master/sql: adding the rrd filename of each DS
This will enable a precomputed mapping table. As all further computation
will be only lookups, maintainance will be easier and paths much more
predictable.

We will also be able to leverage this to group several fields into a
single RRD file for most usages. This allows for some serious
interesting performances improvements.

Note that a plugin cannot always be only one RRD file, only mostly. This
is because the fields on each RRD are specified at creation time. As we
then cannot dynamically add more fields to the common file later,
additional, single-field, RRD files might be needed. (An external
maintainance task is needed if one want to merge them )

Finally that this also enables easy import (with optional updates) of
foreign rrds into munin, as the filename & fieldname is just a value
change. (A full implementation still requires a little more code on
that).