node: Only spoolfetch send 5 samples per service
This enables to not overwhelm the munin-update when there is big backlog
to handle.  Note that this is only a 5x speed increase, so a backlog of
5 hours is fully sent in 1 hour.

This value is _very_ conservative, but a new config value is planned in
the devel branch. Here in stable-2.0, it's a perfectly valid strategy.