Fix iostat on FreeBSD
Current iostat says it shows the graph in MiBps, but it uses iostat's kiB output, so it's not. In this change I've changed the "MB" on the graph to "Bytes" and multiplied the iostat output (it's in kiB as stated by the comments and despite of this handled wrong) by 1024, so munin will correctly SI-scale the graph and numbers (2 MiBps will be 2 MiBps, not 2kiBps [MB] as now).
Also, I've changed the max values from 2 MBps (2000 kiBps) to 8 Gibps (1 GiBps), which is more sane nowadays.
MFM days are long gone.