make proc plugin more careful in /proc/
the proc plugin is using a glob for its grep thats hitting more than
intended. It wants to grep from /proc/$PIDS/stats, but with the current
line it does also take things like /proc/net/stat, and as such leading
to loads of lines like the following in the munin-update.log:

2014/09/16-13:15:26 [587369] Error output from proc:
2014/09/16-13:15:26 [587369] 	   grep: /proc/net/stat: Is a directory

Thats still a hack, but now it limits to anything starting with a number.