mysql: change default database to information_schema
The previous default connected to the mysql database. This required
select privs on the mysql database to connect, however nothing in the mysql
database was actually used by the plugin.

It seems perl DBI insists on a database, or at least is more consistent
to specify one.

Using the information_schema this requires no additional priviledges
from the munin user. It has also been there since at least mysql-5.0.

While we are at it, the permissions for the munin user are documented
and a few more configuration variants are documented.