ttf: remove the TTF fonts
We should not bundle them anymore. That can be handled as a pre-req, and
delegated to the packaging team.

It was introduced in 2006, when tarballs where still a very valid
installation medium. Whereas nowadays, either one prefers to use packaged
stuff, or has ways to deeply automate the retrieval of dependencies.