Update acpi plugin to use the /sys interface.
This is pretty universal now; none of my machines even have
thermal zones in /proc/acpi anymore, but all of them have
them in /sys, which has a much more uniform interface anyway.

As an extra bonus (and the original motivation for the patch),
allows the plugin to understand non-ACPI thermal zones, e.g.
on my ARM SBC, which doesn't have ACPI at all, but has thermal
sensors for both CPU and GPU.

Closes: D:825136