p/postgres_querylength_: don't stack times
The postgres_querylength_ plugin provides two values: The length or
rather age of the oldest transaction and the age of the oldest query. In
the graph these are stacked.

I don't think stacking these values is useful: The sum of these two
values isn't meaningful. In many cases the oldest query is part of the
oldest transaction, so it is counted twice that way, but even if it
isn't you have to mentally subtract the age of the oldest query from the
displayed value for the age of the oldest transaction to get the real
age of the oldest transaction.

I think it is much more useful to display those two values individually:
That way you can immediately see both values and you can also see
whether a long-running transaction consists of a single long-running
query (if both lines overlap) or many small ones (if the query line is
somewhere near the ordinate).

Closes: D:847649