cgi: handle the empty string in CGI arguments
The previous fix is a little too restrictive. And it just adds the parameters
if it is defined. Which should not be the case if it is an empty string.

Note that a simple "if" test has the nasty side-effect of being false if the
value is "0".

A more complete test should be done then.

It might be better to fix the dynazoom call itself, but as we already touched
that area, and failed, let's try to contain this and avoid having the change
spreading epidemically :)