m/Utils: always use munin_read_storable()
Some legacy code used munin_readconf_storable() which was looking for a text
version if the Storable version wasn't present. The text version has been
removed completely, and the munin_readconf_storable() was modified accordingly.

But munin_readconf_storable() didn't get the enhancements of
munin_read_storable(), so this commit removes that old function and remaps the
calls to munin_read_storable()

Closes: #821