Add PostgreSQL 10 support to postgres_xlog
In PostgreSQL 10 the name of the WAL directory was changed so
pg_ls_dir('pg_xlog') no longer works.

Additionally, a new pg_ls_waldir function was added for specifically
listing the WAL directory. This function can be run by any user with the
pg_monitor role, so superuser access is no longer a requirement for this
check.