Makefile: prevent repeated substitution of "Defaults.pm" during "install"
Previously the substitution of paths would happen once during
"make build" and again during "make install".  The latter would mess up
the content of the file, due to the presence of the DESTDIR variable
in the "install" target.

Thus we need to mark a successuful inline substitution of that file
and prevent repetitive operations.