[plugins] remove linux specific df plugin
The linux df plugin uses multigraph for displaying the different devices.
This means that for each device a new graph is generated.
The standard df plugin combines all devices like the df_inode plugin into one single graph.
I think this is a more compact way.

p.s.: currently the multigraph support is also broken, cause a dot instead of an underscore is used as seperator.