Plugin cpuspeed: add support for 'intel_pstate' driver
This changeset introduces optional support the 'intel_pstate' driver for
accessing the speed of each CPU.

Previously 'acpi-cpufreq' kernel module was the only data source of this
plugin.  Sadly Intel-based CPUs started to use the 'intel_pstate' driver
instead of 'acpi-cpufreq' since Linux 3.9.  The 'intel_pstate' driver
does not provide the statistical data that is available via
'acpi-cpufreq', thus users of 'intel_pstate' will only see the
"current" (not the avarage) speed of each CPU at the moment of the data
retrieval.  This may result in discrete speed steps instead of a smooth
average (depending on the CPU type) and most probably does not reflect
the real CPU speed usage.

Details:
* https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=717287
* https://bugzilla.kernel.org/show_bug.cgi?id=57941