munin-async.postinst: fix handling of a missing "munin-async" group
do not try to create the user again, but add it to the group instead

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=531021#17 for
the origin of this workaround.
See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=910684#89 for
an error message related to the issue fixed by this commit.