include empty directory /var/lib/munin-async in "munin-node" instead of "munin"
Previously the directory /var/lib/munin-async was part of the packages
"munin" (accidentally) and "munin-async" (on purpose). Instead it should
have been part of the "munin-node" package. The reasoning for each
package follows:
* munin: the directory is not used by any master program
* munin-node: the directory is required by "munin-sched"
* munin-async: the directory is created via "adduser" in the postinst
  script and is used by "munin-async" and "munin-asyncd"

This commit removes the directory from the package "munin" and adds it
to the package "munin-node" instead.

Sadly the requirement for this directory due to "munin-sched" (as part
of "munin-node") conflicts with the creation of that directory via
"adduser" in "munin-async". Thus a warning during "adduser" was emitted
before and will be emitted in the future.