Tests: temporarily allow 'munindoc' to write to stderr
munindoc (as of munin v2.0.42) emits text to stderr. This is considered
to be a failure by sharness.
Thus we silence this output for now and add a test for bailing out, as
soon as the issue is fixed.

The upstream issue is fixed in 3a629ac5f815365744579f02ee05f931557c46ec.