Plugin amavis: only enable via "autoconf" if amavisd is installed
Previously "autoconf" failed due to the test for 'logtail' (thus
requiring manual configuration).  This issue was fixed in 6b43263fb5c.

Thus "autoconf" is relevant from now on.  Currently it succeeds on
every reasonably configured host with a mail log file and with "logtail"
installed.
The additional check for "amavisd" (or "amavisd-new") allows to
distinguish between a common mail host and a host suitable for this
plugin.

Thanks to barajus for proposing this check:
 https://github.com/munin-monitoring/munin/pull/986