Tests: allow to run in an autopkgtest environment lacking a FQDN
An autopkgtest environment (e.g. via LXC) may lack a domain and thus
provide only a hostname (e.g. "autopkgtest-lxc-hmraig").
Since it is hard to detect such an environment, where a domain is
missing, we rely on the hostname following such a name template.
This is not a clean solution, but it should hardly cause
misdetections.