Attempt to fix the autopkgtest failure experienced on ci.d.n
The test for the default user executing the munin plugins currently
fails on the infrastructure of ci.debian.net.  A possible cause is the
restart of the munin-node service, which may return (due to
"Type=forking") before the daemon is able to respond to requests. Thus
a bit of delay after the restart may help to fix the test failure.

Signed-off-by: Holger Levsen <holger@layer-acht.org>