Remove master configuration setting "use_default_node"
The setting was introduced in 6adf38ae61a.

But somehow it partially lost its meaning in the merge commit
c8228b8e5bb (Merge branch 'sql' into devel).
Afterwards it was only used for for breaking the node communication
in case of a missing "host" attribute (which can never be empty as
it is an integral part of the node specification in the master
configuration).  Thus it is not useful anymore.

The purpose of this setting was later replaced with
"use_node_name ignore" in dd4cbb501ca.