Remove "ignore" value for "use_node_name"
The additional value "ignore" was added in dd4cbb501ca7152.

It caused the following "list" commands to be sent to the node:
* "yes" -> "list $self->{node_name}"
* "ignore" -> "list "
* "no" -> "list $self->{host}"
* any other value -> "list $self->{host}"

The above "$self->{node_name}" is the name advertised by the node during
the opening of the connection.
"$self->{host}" is the name of the node section in the master
configuration.

The new behaviour is the following:
* "yes" -> "list"
* "ignore" -> "list"
* "no" -> "list $self->{host}"
* any other value -> "list $self->{host}"

This behaviour has the same effect as before, as the request for "list"
(without a specific node name) is handley by munin-node exactly, as if
its "node_name" is supplied.