Plugin apt_all: handle configured list of "releases" properly
Previously the "split" of releases was interpreted as a scalar and thus
resulted in only one item (containing the number of configured
releases). Thus the plugin emitted the a field like "hold_1".

Now the "split" is forced to be interpreted in list context. Thus the
release strings are used for the fields.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=715141