munin-cgi-html: avoid accidental permanent path manipulation
In case munin-cgi-html is exeucuted in a permanent context (e.g. via
fcgid or as a systemd socket service) it accidentally manipulated a path
variable permanently.

The issue can be reproduced with the following steps:
* request a service page (e.g. "/munin/GROUP/HOST/df.html")
    * links to "dynazoom" are valid
* request a time-based category comparison (e.g. "/munin/disk-day.html")
* request a service page (e.g. "/munin/GROUP/HOST/df.html")
    * links to "dynazoom" lack the relative "../../" prefix

Thus we need to restore the previous path settings after emitting a
top-level page (e.g. a time-based category comparison).

This fix is a bit hacky (just like the problematic path manipulation
itself).  But given that `munin-2.0` is effectively in maintenance-only
mode, this rather cumbersome minimal fix should be acceptable.

Closes: #460

Thanks to gnuheidix (https://github.com/gnuheidix) for preparing an
excellent minimal procedure for reproducing this issue.