tests: verify name munging of remote plugins via async
Plugin names may contain characters that are forbidden for multigraph
names.  Thus we need to munge these names before storing the data.
These updated tests verify that data to be stored in the spool directory
is stored under a proper filename (without forbidden characters).
Additionally it checks whether written data is accessible for the spool
reader under the munged name.