Plugin open_files: remove "max" field from graph
The maximum number of open files recently (systemd version >= 240)
jumped to a large value (18 quintillion).  This changed the scaling of
the graph significantly and obscured the visibility of the number of
open files.

The warning and critical thresholds for the number of open files are
not affected by this change.

Contributed by Gerald Turner
see https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=928211