Fix the reversal of path manipulation
The change 2fde4da9a1 worked for `html_category cgi`.  But it broke the
intended path manipulation for `cron`.
The problem for `cron` was caused by the late resolution of the template
output (happening after the path manipulation was restored to its
original value).
Now the template output is stored in a variable before the path
values are restored.