Add image path manipulation for "problems" graphs
The issue is only relevant for "html_strategy cron".

The critical/warnings/unknowns graphs were not rendered due to broken
image links, since they did not take the path level of the source page
into account.
This problem is visible since 2fde4da9a17d.
Before that commit, the underlying problem was hidden by the missing
path manipulation restoring for the category templates (introduced in
2fde4da9a17d).
As soon as this path manipulation was properly restored in that commit,
the graph links on the "problems" page were not manipulated anymore and
thus wrong.

Conclusion: add the (dirty) path manipulation (and restoring) to the
"problems" page.

Now all non-regular graph overviews use properly manipulated (and
restored) graph links:
* categories
* time-based comparisons on the group level
* problems