sample config: improve notification mail subject
The previous notification mail subject did not contain any helpful
information.  Thus the receiver needed to open the mail in order to gain
insights into the relevance of the issue.

The new example configuration adds the most relevant state change (e.g.
"CRIITCAL") as well as the host and the plugin name to the mail subject.

Closes: #1085