Plugin sensors_: add support for voltage output with unit prefix (mV)
Previously the output format of "sensors" used only "V":

  Vcore:         1.02 V  (min =  +0.00 V, max =  +1.74 V)

Recently the values started to be optionally formatted with a unit
prefix:

  Vcore:       880.00 mV (min =  +0.00 V, max =  +1.74 V)

THe plugin should really use "sensors -u" instead of "sensors", as this
would provide a stable output format.  But the "sensors_" plugin is
not part of munin 3.0 anymore.  Thus a quick workaround is sufficient.

Closes: d#943577