Tests: simplify handling of different init systems
It seems to be possible now to install sysvinit-core via autopkgtest
without manual confirmation.  Thus we can install the package via the
test description.  The only required additional work is a manual check,
whether the new init system is already active.  We solve this by simply
forcing a reboot once.