munin-run: apply systemd properties configured for "munin-node.service"
The "munin-node" service process may be restricted by systemd
properties.  In contrast, "munin-run" was executing the plugin with
unrestricted root permissions.  This was a source of confusion for
developers of munin plugins.

From now on "munin-run" tries to detect the locally configured
systemd properties of the "munin-node" service
("systemctl show munin-node") and applies these to "munin-run" itself
via "systemd-run".