Respect custom color for warning thresholds
This change finalizes the intent of commit 0cb219a11e1.
That commit only fixed the color selection for warning thresholds in
case of graphs with a negative field.
Now the configured (instead of the default color) is also used for
graphs without a negative value.

See #1135 and 0cb219a11e1.