munin-run: work around systemd's serialization of EnvironmentFile
The output of "systemd show munin-node" may contain the following lines:
1)  EnvironmentFiles=/etc/default/ssh (ignore_errors=yes)
2)  EnvironmentFile=/etc/default/ssh (ignore_errors=yes)
The first line is emitted by systemd v242 or later.  The second line is
emitted by earlier versions of systemd.
The first line is ignored by "munin-run" since the plural form of
"EnvironmentFile" is not whitelisted (thus it is ignored, which is also
not desirable).  But the second line is treated by munin-run as a valid
property, even though "systemd-run" refuses to parse the "ignore_errors"
suffix.  Thus "munin-run" failed on hosts with systemd before v242.

Now we simply ignore such ill-formatted lines and wait for systemd to
fix this issue.  The ignored environment files are rarely used.

See https://github.com/systemd/systemd/issues/14723.