Implement "sd_notify" interface for munin-node
The "sd_notify" interface [1] allows a service to announce its readiness
via a simple text-based protocol.  This avoids timing issues during
service bootup.

The "sd_notify" interface is supported by:
  * systemd: see "Type=Notify" in "man systemd.exec"
  * start-stop-daemon: see "--notify-await" in "man start-stop-daemon"
(and maybe others)

It does not hurt, if the caller does not support this interface.
In this case the notification is simply not sent.

[1] https://www.freedesktop.org/software/systemd/man/sd_notify.html