munin-node: allow to force "foreground" mode via commandline arguments
The recently supported "sd_notify" interface can signal the readiness of
a process to its caller.
Currently "munin-node" forks into background, if the configuration file
contains one of the settings "background 1" or "setsid 1".  Sadly the
fork currently happens before the service is ready (and the signal is
sent).

Both systemd and start-stop-daemon assume a non-forking process to be
tracked via "sd_notify".  Thus we allow packagers of munin to override
existing "background" and "setsid" settings in the configuration file
via a new commandline switch ("--foreground").