d/tests: wait for the munin init script before starting master tests
"autopkgtest-virt-lxc" waits for a numeric runlevel being reached.  But this
only verifies, that "init" is now targeting the runlevel - but it is not
necessarily reached, yet.
Thus we wait until the sysvinit script (/etc/init.d/rc) is finished.
Otherwise the first execution of "munin-update" or the first access of
the "munin-node" service may fail.

This waiting condition can be removed, as soon as d#953655 is resolved.

Closes: #949887