Proc plugin: Allow monitoring of processes with empty cmdline
If /proc/pid/cmdline contains nothing, the plugin would not use the counters.

cmdline should only be needed if one have specified command line arguments,
and some processes (like kswapd0 for example) do not have any contents in its
cmdline.

This patch allows kswapd0 and friends to be monitored with this plugin.